+++ test/verify/test_verify.py
@@ -122,7 +122,8 @@ class TestVerify:
             # Create difficulty rule.
             difficulty_rule = DifficultyRule(
                 org_group_id=org_group.id,
-                taxon_id=taxon.id,
+                organism_key=taxon.organism_key,
+                taxon=taxon.name,
                 difficulty_code_id=difficulty_code.id
             )
             db.add(difficulty_rule)
@@ -164,7 +165,8 @@ class TestVerify:
             # Create tenkm rule for org_group and taxon.
             rule1 = TenkmRule(
                 org_group_id=org_group.id,
-                taxon_id=taxon.id,
+                organism_key=taxon.organism_key,
+                taxon=taxon.name,
                 km100='TL',
                 km10='14 15 16',
                 coord_system='OSGB'